refactor: change api for worksheets
The main issue from before is that all parsing of a worksheet happened on `workbook.worksheet()`,
but if you wanted to initiate the parsing in different ways for a worksheet, that wasn't possible.

These changes begin to split up when the parsing is done to the level of choosing how you want
to iterate over the cells. More directly, this allows for "specialization" impls tied to the type
of workbook the worksheet is in. The hope is that this granular approach will still allow common
ground in which can act as a glue between the various spreadsheet formats.

Known issues: There is a lifetime issue with the workbook function.